% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apus-package.R
\docType{package}
\name{apus-package}
\alias{apus}
\alias{apus-package}
\title{apus: Advises an optimal fertilizer mix for a farm}
\description{
apus optimizes the fertilizer mix (type and dose) for a farm based on the agronomic nutrient requirements, legal norms, crop-dose response curves, available fertilizers in order to find the best economic result within the provided agronomic and environmental boundaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AgroCares/apus}
  \item Report bugs at \url{https://github.com/AgroCares/apus/issues}
}

}
\author{
\strong{Maintainer}: Sven Verweij \email{sven.verweij@nmi-agro.nl} (\href{https://orcid.org/0000-0002-5573-3952}{ORCID})

Authors:
\itemize{
  \item Gerard Ros \email{gerard.ros@nmi-agro.nl}
}

Other contributors:
\itemize{
  \item Nutriënten Managment Instituut \email{nmi@nmi-agro.nl} [copyright holder]
}

}
\keyword{internal}
