% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apus.R
\name{Apus}
\alias{Apus}
\title{R6 Class Representing a farm to optimize the fertilizer mix for}
\description{
A farm has a name
}
\details{
A farm needs to have fields, fertilizer etc
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{farm_name}}{(character) Name of the farm}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Apus-new}{\code{Apus$new()}}
\item \href{#method-Apus-clone}{\code{Apus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Apus-new"></a>}}
\if{latex}{\out{\hypertarget{method-Apus-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apus$new(farm_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{farm_name}}{(character) The name of the farm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Apus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Apus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
