% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ApusModel}
\alias{ApusModel}
\title{R6 Class Representing a model that can be used to optimize}
\description{
A model has
}
\details{
A model requires
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{(character) Name of the farm}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApusModel-new}{\code{ApusModel$new()}}
\item \href{#method-ApusModel-trainModel}{\code{ApusModel$trainModel()}}
\item \href{#method-ApusModel-clone}{\code{ApusModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApusModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApusModel-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApusModel$new(
  cultivations = apus::cultivations,
  fertilizers = apus::fertilizers,
  fines = apus::fines
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cultivations}}{(data.table) The list of cultivations available to train the model for. Should have similiar structure}

\item{\code{fertilizers}}{(data.table)}

\item{\code{fines}}{(data.table)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApusModel-trainModel"></a>}}
\if{latex}{\out{\hypertarget{method-ApusModel-trainModel}{}}}
\subsection{Method \code{trainModel()}}{
Train an apus model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApusModel$trainModel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApusModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApusModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApusModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
