% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{createApusModel}
\alias{createApusModel}
\title{A torch model for apus}
\usage{
createApusModel(device)
}
\arguments{
\item{device}{(character)}

\item{farms}{(data.table)}
}
\description{
Creates a torch model to be used for apus
}
