% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cultivations}
\alias{cultivations}
\title{Cultivation table}
\format{
\describe{
\item{b_lu}{The landuse (cultivation) identifier code}
\item{b_lu_name}{The name of the cultivation}
\item{b_lu_yield}{The expected yield in kg/ha under optimal conditions}
\item{b_lu_price}{The price of yielded cultivation per kg}
}
}
\usage{
cultivations
}
\description{
A data.table with the cultivations available
}
\keyword{datasets}
