% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertilizers}
\alias{fertilizers}
\title{Fertilizers table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 47 rows and 52 columns.
}
\usage{
fertilizers
}
\description{
A data.table with the fertilizers available
}
\keyword{datasets}
