% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fines}
\alias{fines}
\title{Fines table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 2 columns.
}
\usage{
fines
}
\description{
A data.table with the fines available
}
\keyword{datasets}
