% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Parameters table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11 rows and 13 columns.
}
\usage{
parameters
}
\description{
A data.table with the parameters used in this package and their definition
}
\keyword{datasets}
