% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apus-package.R
\docType{package}
\name{apus-package}
\alias{apus}
\alias{apus-package}
\title{apus: Advises an optimal fertilizer mix for a farm}
\description{
apus optimizes the fertilizer mix for a farm based on the nutrient requirements, legal norms, cultivation response curves, available fertilizers and tries to archieve the best economic result within the provided boundaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AgroCares/apus}
  \item Report bugs at \url{https://github.com/AgroCares/apus/issues}
}

}
\author{
\strong{Maintainer}: Sven Verweij \email{sven.verweij@nmi-agro.nl} (\href{https://orcid.org/0000-0002-5573-3952}{ORCID})

Other contributors:
\itemize{
  \item Nutriënten Managment Instituut \email{nmi@nmi-agro.nl} [copyright holder]
}

}
\keyword{internal}
