% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apus.R
\name{Apus}
\alias{Apus}
\title{R6 Class Representing a farm to optimize the fertilizer mix for}
\description{
A farm has a name
}
\details{
A farm needs to have fields, fertilizer etc
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{farm_name}}{(character) Name of the farm}

\item{\code{fields}}{(data.table) A table with the fields and their properties}

\item{\code{cultivations}}{(data.table) A table with the properties of the cultivations}

\item{\code{fertilizers}}{(data.table) A table with the properties of the fertilizers}

\item{\code{fines}}{(data.table) A table with the properties of the fines}

\item{\code{model}}{(nn_module) The model to use for optimization of fertilizer choice}

\item{\code{device}}{(character) Whether to run the model on \code{cpu} or \code{cuda}}

\item{\code{fields_max}}{(integer) Maximum number of fields for a farm}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Apus-new}{\code{Apus$new()}}
\item \href{#method-Apus-addField}{\code{Apus$addField()}}
\item \href{#method-Apus-trainModel}{\code{Apus$trainModel()}}
\item \href{#method-Apus-optimizeFertilizerChoice}{\code{Apus$optimizeFertilizerChoice()}}
\item \href{#method-Apus-clone}{\code{Apus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Apus-new"></a>}}
\if{latex}{\out{\hypertarget{method-Apus-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apus$new(farm_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{farm_name}}{(character) The name of the farm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Apus-addField"></a>}}
\if{latex}{\out{\hypertarget{method-Apus-addField}{}}}
\subsection{Method \code{addField()}}{
Add a field to the apus object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apus$addField(
  b_id_field,
  b_area,
  b_lu,
  d_n_req = NA,
  d_p_req = NA,
  d_k_req = NA,
  d_n_norm = NA,
  d_n_norm_man = NA,
  d_p_norm = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b_id_field}}{(character) ID or (unique) name of the field}

\item{\code{b_area}}{(number) The area of the field (m^2)}

\item{\code{b_lu}}{(character) The cultivation code for this field}

\item{\code{d_n_req}}{(number) The required amount of Nitrogen for this field (kg N / ha)}

\item{\code{d_p_req}}{(number) The required amount of Phosphate for this field (kg P2O5 / ha)}

\item{\code{d_k_req}}{(number) The required amount of Potassium for this field (kg K2O / ha)}

\item{\code{d_n_norm}}{(number) The legal limit for workable Nitrogen (kg N / ha)}

\item{\code{d_n_norm_man}}{(number) The legal limit for total Nitrogen from manure (kg N / ha}

\item{\code{d_p_norm}}{(number) The legal limit for Phosphate (kg P2O5 / ha)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Apus-trainModel"></a>}}
\if{latex}{\out{\hypertarget{method-Apus-trainModel}{}}}
\subsection{Method \code{trainModel()}}{
Train a model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apus$trainModel(width = 12, layers = 1, epochs = 3, device = "cpu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{(integer)}

\item{\code{layers}}{(integer)}

\item{\code{epochs}}{(integer)}

\item{\code{device}}{(character)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Apus-optimizeFertilizerChoice"></a>}}
\if{latex}{\out{\hypertarget{method-Apus-optimizeFertilizerChoice}{}}}
\subsection{Method \code{optimizeFertilizerChoice()}}{
Optimize fertilizer choice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apus$optimizeFertilizerChoice()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Apus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Apus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
