% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{createApusDataset}
\alias{createApusDataset}
\title{A torch dataset for apus}
\usage{
createApusDataset(farms = NULL, device)
}
\arguments{
\item{farms}{(data.table)}

\item{device}{(character)}
}
\description{
Creates a torch dateset to be used for apus model
}
