% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{createApusModel}
\alias{createApusModel}
\title{A torch model for apus}
\usage{
createApusModel(dataset, width = 12, depth = 1, device)
}
\arguments{
\item{dataset}{(ApusDataset)}

\item{width}{(integer)}

\item{depth}{(integer)}

\item{device}{(character)}
}
\description{
Creates a torch model to be used for apus
}
